Update 5
This Update entails the fixing of the animation at the homescreen using javascript, we aso added a loading animation so that whenever the page is not loaded, we wouldn't get something unclean as the product, but we would get a loading animation until the loading is done. The style for the loading is in styles.css so it could be used universally as long as you add the div and the script. Added some images to the slideshow, tesla gif needs to be better thought of: too short.